+++ Turculescu_Samuel/Week_5/Domain/Models/CartItem.cs
@@ -10,10 +10,10 @@ namespace Domain.Models
         public string Name { get; set; }
         public string SpecialRequests { get; set; }
         public uint Quantity { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public MenuItem MenuItem { get; set; }
 
-        public CartItem(string name, string specialRequests, uint quantity, double price, MenuItem menuItem)
+        public CartItem(string name, string specialRequests, uint quantity, decimal price, MenuItem menuItem)
         {
             Name = name;   
             SpecialRequests = specialRequests;
